+++ src/components/MultiSelect/MultiSelect.stories.tsx
@@ -318,11 +318,13 @@ export const withControls: Story = {
   args: {
     buttonText: "MultiSelect",
     closeOnBlur: true,
+    defaultItemsVisible: undefined,
     isBlockElement: true,
     isDefaultOpen: false,
     isSearchable: true,
     items: withItems,
     listOverflow: "scroll",
+    onClear: undefined,
     onChange: undefined,
     onMixedStateChange: undefined,
     selectedItems: undefined,
@@ -659,6 +661,7 @@ const MultiSelectWithControlsStory = (args) => {
     />
   );
 };
+MultiSelectWithControlsStory.displayName = "MultiSelectWithControlsStory";
 
 const MultiSelectStory = ({
   id,
@@ -719,6 +722,7 @@ const MultiSelectStory = ({
     />
   );
 };
+MultiSelectStory.displayName = "MultiSelectStory";
 
 // TODO: Replace with MultiSelectGroup once that component is done.
 const MultiSelectGroupStory = ({ items }: Partial<MultiSelectProps>) => {
@@ -816,3 +820,4 @@ const MultiSelectGroupStory = ({ items }: Partial<MultiSelectProps>) => {
     </HStack>
   );
 };
+MultiSelectGroupStory.displayName = "MultiSelectGroupStory";